requirements.txt
qdrant-client
sentence-transformers
Pillow
Flask
gunicorn